refactor: showDetail 공통 함수로 개선 - 콜백: getUrl로 이동할 주소값을 받아 id값을 넣어줌
