--- admin/pages/components/order/showDetails.js
@@ -1,11 +1,0 @@
-export function showDetails(button, orderId) {
-  if (!button) return;
-  if (orderId == null) {
-    console.error("orderId가 없습니다.");
-    return;
-  }
-
-  button.addEventListener("click", () => {
-    window.location.href = `/admin/pages/order-details.html?orderId=${orderId}`;
-  });
-}
