--- admin/pages/components/user/showDetails.js
@@ -1,11 +1,0 @@
-export function showDetails(button, userId) {
-  if (!button) return;
-  if (userId == null) {
-    console.error("userId가 없습니다.");
-    return;
-  }
-
-  button.addEventListener("click", () => {
-    window.location.href = `/admin/orders/${userId}`;
-  });
-}
